# backend/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/reports": {
      "post": {
        "operationId": "ReportController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateReportDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Report created successfully"
          },
          "400": {
            "description": "Bad request - validation failed"
          },
          "401": {
            "description": "Unauthorized"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Create a new report",
        "tags": [
          "reports"
        ]
      },
      "get": {
        "operationId": "ReportController_findAll",
        "parameters": [
          {
            "name": "type",
            "required": false,
            "in": "query",
            "description": "Filter by report type",
            "schema": {
              "type": "string",
              "enum": [
                "INCIDENT",
                "INCIDENT_SUMMARY",
                "TASK",
                "TASK_PROGRESS",
                "SURVEY",
                "SURVEY_RESULTS",
                "MONTHLY",
                "MONTHLY_SUMMARY",
                "CUSTOM"
              ]
            }
          },
          {
            "name": "status",
            "required": false,
            "in": "query",
            "description": "Filter by report status",
            "schema": {
              "type": "string",
              "enum": [
                "DRAFT",
                "GENERATING",
                "READY",
                "ERROR",
                "SUBMITTED",
                "UNDER_REVIEW",
                "REVISION_REQUIRED",
                "APPROVED",
                "REJECTED"
              ]
            }
          },
          {
            "name": "incidentId",
            "required": false,
            "in": "query",
            "description": "Filter by incident ID",
            "schema": {
              "example": "123e4567-e89b-12d3-a456-426614174000",
              "type": "string"
            }
          },
          {
            "name": "authorId",
            "required": false,
            "in": "query",
            "description": "Filter by author ID",
            "schema": {
              "example": "123e4567-e89b-12d3-a456-426614174000",
              "type": "string"
            }
          },
          {
            "name": "periodStartFrom",
            "required": false,
            "in": "query",
            "description": "Filter by period start date (from)",
            "schema": {
              "example": "2025-01-01T00:00:00Z",
              "type": "string"
            }
          },
          {
            "name": "periodStartTo",
            "required": false,
            "in": "query",
            "description": "Filter by period start date (to)",
            "schema": {
              "example": "2025-01-31T23:59:59Z",
              "type": "string"
            }
          },
          {
            "name": "createdFrom",
            "required": false,
            "in": "query",
            "description": "Filter by creation date (from)",
            "schema": {
              "example": "2025-01-01T00:00:00Z",
              "type": "string"
            }
          },
          {
            "name": "createdTo",
            "required": false,
            "in": "query",
            "description": "Filter by creation date (to)",
            "schema": {
              "example": "2025-01-31T23:59:59Z",
              "type": "string"
            }
          },
          {
            "name": "page",
            "required": false,
            "in": "query",
            "description": "Page number for pagination",
            "schema": {
              "default": 1,
              "example": 1,
              "type": "number"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "description": "Number of items per page",
            "schema": {
              "default": 10,
              "example": 10,
              "type": "number"
            }
          },
          {
            "name": "sortBy",
            "required": false,
            "in": "query",
            "description": "Sort field",
            "schema": {
              "default": "createdAt",
              "example": "createdAt",
              "type": "string"
            }
          },
          {
            "name": "sortOrder",
            "required": false,
            "in": "query",
            "description": "Sort order",
            "schema": {
              "default": "desc",
              "example": "desc",
              "type": "string",
              "enum": [
                "asc",
                "desc"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Reports retrieved successfully"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get all reports with filtering and pagination",
        "tags": [
          "reports"
        ]
      }
    },
    "/reports/statistics": {
      "get": {
        "operationId": "ReportController_getStatistics",
        "parameters": [
          {
            "name": "periodStart",
            "required": false,
            "in": "query",
            "description": "Start date for statistics",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "periodEnd",
            "required": false,
            "in": "query",
            "description": "End date for statistics",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "type",
            "required": false,
            "in": "query",
            "description": "Filter by report type",
            "schema": {
              "enum": [
                "INCIDENT",
                "INCIDENT_SUMMARY",
                "TASK",
                "TASK_PROGRESS",
                "SURVEY",
                "SURVEY_RESULTS",
                "MONTHLY",
                "MONTHLY_SUMMARY",
                "CUSTOM"
              ],
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Statistics retrieved successfully"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get report statistics",
        "tags": [
          "reports"
        ]
      }
    },
    "/reports/{id}": {
      "get": {
        "operationId": "ReportController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Report ID",
            "schema": {
              "example": "123e4567-e89b-12d3-a456-426614174000",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Report retrieved successfully"
          },
          "404": {
            "description": "Report not found"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get a report by ID",
        "tags": [
          "reports"
        ]
      },
      "patch": {
        "operationId": "ReportController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Report ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateReportDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Report updated successfully"
          },
          "403": {
            "description": "Forbidden - can only update own reports"
          },
          "404": {
            "description": "Report not found"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Update a report",
        "tags": [
          "reports"
        ]
      },
      "delete": {
        "operationId": "ReportController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Report ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Report deleted successfully"
          },
          "403": {
            "description": "Forbidden - can only delete own reports"
          },
          "404": {
            "description": "Report not found"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Delete a report",
        "tags": [
          "reports"
        ]
      }
    },
    "/reports/{id}/submit": {
      "post": {
        "operationId": "ReportControllerExtension_submitFullReport",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Report submitted successfully"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Submit full report for review",
        "tags": [
          "reports"
        ]
      }
    },
    "/reports/{id}/review": {
      "post": {
        "operationId": "ReportController_review",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Report ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReviewReportDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Report reviewed successfully"
          },
          "400": {
            "description": "Bad request - report cannot be reviewed"
          },
          "403": {
            "description": "Forbidden - insufficient permissions"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Review a report (SUPERVISOR, EXECUTIVE, ADMIN only)",
        "tags": [
          "reports"
        ]
      }
    },
    "/reports/{id}/generate-pdf": {
      "post": {
        "operationId": "ReportController_generatePdf",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Report ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GeneratePdfDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "PDF generation started successfully"
          },
          "404": {
            "description": "Report not found"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Generate PDF for a report",
        "tags": [
          "reports"
        ]
      }
    },
    "/reports/{id}/pdf": {
      "get": {
        "operationId": "ReportController_downloadPdf",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Report ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "PDF file",
            "content": {
              "application/pdf": {}
            }
          },
          "404": {
            "description": "Report or PDF not found"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Download PDF of a report",
        "tags": [
          "reports"
        ]
      }
    },
...
  },
  "info": {
    "title": "Guardian Route API",
    "description": "\n      Guardian Route - ระบบบริหารจัดการภัยพิบัติ\n      ตำบลเวียง อำเภอฝาง จังหวัดเชียงใหม่\n      \n      ## Authentication\n      Most endpoints require JWT Bearer token authentication.\n      Use the /api/auth/login endpoint to obtain a token.\n      \n      ## Roles\n      - DEVELOPER: Full system access\n      - ADMIN: System administration\n      - SUPERVISOR: Command and control\n      - EXECUTIVE: Strategic oversight\n      - FIELD_OFFICER: Field operations\n      \n      ## Base URL\n      - Development: http://localhost:3001/api\n      - Production: https://api.guardian-route.example.com/api\n    ",
    "version": "1.0.0",
    "contact": {
      "name": "Guardian Route Team",
      "url": "https://github.com/jetci/Guardian-Route",
      "email": "support@guardian-route.example.com"
    },
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/licenses/MIT"
    }
  },
...
  "servers": [
    {
      "url": "http://localhost:3001",
      "description": "Development server"
    },
    {
      "url": "https://api.guardian-route.example.com",
      "description": "Production server"
    }
  ],
  "components": {
    "securitySchemes": {
      "JWT-auth": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http",
        "name": "Authorization",
        "description": "Enter JWT token",
        "in": "header"
      }
    },
    "schemas": {
...
      "CreateReportDto": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string",
            "description": "Type of report",
            "enum": [
              "INCIDENT",
              "INCIDENT_SUMMARY",
              "TASK",
              "TASK_PROGRESS",
              "SURVEY",
              "SURVEY_RESULTS",
              "MONTHLY",
              "MONTHLY_SUMMARY",
              "CUSTOM"
            ],
            "example": "INCIDENT"
          },
          "title": {
            "type": "string",
            "description": "Report title",
            "example": "Flood Damage Assessment - Village 1"
          },
          "summary": {
            "type": "string",
            "description": "Report summary",
            "example": "Summary of flood damage in Village 1 affecting 15 households"
          },
          "details": {
            "type": "object",
            "description": "Structured report data",
            "example": {
              "findings": [
                "Severe flooding in low-lying areas"
              ],
              "recommendations": [
                "Improve drainage system"
              ]
            }
          },
          "totalDamageEstimate": {
            "type": "number",
            "description": "Total damage estimate in THB",
            "example": 500000
          },
          "affectedHouseholds": {
            "type": "number",
            "description": "Number of affected households",
            "example": 15
          },
          "affectedPersons": {
            "type": "number",
            "description": "Number of affected persons",
            "example": 45
          },
          "aiAnalysis": {
            "type": "object",
            "description": "AI analysis results from Gemini API",
            "example": {
              "severity": "high",
              "recommendations": [
                "Immediate evacuation required"
              ]
            }
          },
          "photoUrls": {
            "description": "Array of photo URLs",
            "example": [
              "https://example.com/photo1.jpg",
              "https://example.com/photo2.jpg"
            ],
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "templateId": {
            "type": "string",
            "description": "Template ID for custom reports",
            "example": "123e4567-e89b-12d3-a456-426614174000"
          },
          "periodStart": {
            "type": "string",
            "description": "Period start date (for monthly/custom reports)",
            "example": "2025-01-01T00:00:00Z"
          },
          "periodEnd": {
            "type": "string",
            "description": "Period end date (for monthly/custom reports)",
            "example": "2025-01-31T23:59:59Z"
          },
          "metadata": {
            "type": "object",
            "description": "Additional metadata for report generation",
            "example": {
              "filters": {
                "village": "1",
                "status": "RESOLVED"
              },
              "parameters": {
                "includeImages": true
              }
            }
          },
          "incidentId": {
            "type": "string",
            "description": "Incident ID (required for INCIDENT reports)",
            "example": "123e4567-e89b-12d3-a456-426614174000"
          },
          "status": {
            "type": "string",
            "description": "Initial status of the report",
            "enum": [
              "DRAFT",
              "GENERATING",
              "READY",
              "ERROR",
              "SUBMITTED",
              "UNDER_REVIEW",
              "REVISION_REQUIRED",
              "APPROVED",
              "REJECTED"
            ],
            "example": "DRAFT",
            "default": "DRAFT"
          }
        },
        "required": [
          "type",
          "title"
        ]
      },
      "UpdateReportDto": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string",
            "description": "Type of report",
            "enum": [
              "INCIDENT",
              "INCIDENT_SUMMARY",
              "TASK",
              "TASK_PROGRESS",
              "SURVEY",
              "SURVEY_RESULTS",
              "MONTHLY",
              "MONTHLY_SUMMARY",
              "CUSTOM"
            ],
            "example": "INCIDENT"
          },
          "title": {
            "type": "string",
            "description": "Report title",
            "example": "Flood Damage Assessment - Village 1"
          },
          "summary": {
            "type": "string",
            "description": "Report summary",
            "example": "Summary of flood damage in Village 1 affecting 15 households"
          },
          "details": {
            "type": "object",
            "description": "Structured report data",
            "example": {
              "findings": [
                "Severe flooding in low-lying areas"
              ],
              "recommendations": [
                "Improve drainage system"
              ]
            }
          },
          "totalDamageEstimate": {
            "type": "number",
            "description": "Total damage estimate in THB",
            "example": 500000
          },
          "affectedHouseholds": {
            "type": "number",
            "description": "Number of affected households",
            "example": 15
          },
          "affectedPersons": {
            "type": "number",
            "description": "Number of affected persons",
            "example": 45
          },
          "aiAnalysis": {
            "type": "object",
            "description": "AI analysis results from Gemini API",
            "example": {
              "severity": "high",
              "recommendations": [
                "Immediate evacuation required"
              ]
            }
          },
          "photoUrls": {
            "description": "Array of photo URLs",
            "example": [
              "https://example.com/photo1.jpg",
              "https://example.com/photo2.jpg"
            ],
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "templateId": {
            "type": "string",
            "description": "Template ID for custom reports",
            "example": "123e4567-e89b-12d3-a456-426614174000"
          },
          "periodStart": {
            "type": "string",
            "description": "Period start date (for monthly/custom reports)",
            "example": "2025-01-01T00:00:00Z"
          },
          "periodEnd": {
            "type": "string",
            "description": "Period end date (for monthly/custom reports)",
            "example": "2025-01-31T23:59:59Z"
          },
          "metadata": {
            "type": "object",
            "description": "Additional metadata for report generation",
            "example": {
              "filters": {
                "village": "1",
                "status": "RESOLVED"
              },
              "parameters": {
                "includeImages": true
              }
            }
          },
          "incidentId": {
            "type": "string",
            "description": "Incident ID (required for INCIDENT reports)",
            "example": "123e4567-e89b-12d3-a456-426614174000"
          },
          "status": {
            "type": "string",
            "description": "Initial status of the report",
            "enum": [
              "DRAFT",
              "GENERATING",
              "READY",
              "ERROR",
              "SUBMITTED",
              "UNDER_REVIEW",
              "REVISION_REQUIRED",
              "APPROVED",
              "REJECTED"
            ],
            "example": "DRAFT",
            "default": "DRAFT"
          }
        }
      },
...
      "ReviewReportDto": {
        "type": "object",
        "properties": {
          "status": {
            "type": "string",
            "description": "New status after review",
            "enum": [
              "APPROVED",
              "REJECTED",
              "REVISION_REQUIRED"
            ],
            "example": "APPROVED"
          },
          "reviewNotes": {
            "type": "string",
            "description": "Review notes",
            "example": "Report approved with minor comments"
          }
        },
        "required": [
          "status"
        ]
      },
      "GeneratePdfDto": {
        "type": "object",
        "properties": {
          "options": {
            "type": "object",
            "description": "PDF generation options",
            "example": {
              "includeImages": true,
              "includeCharts": true,
              "pageSize": "A4",
              "orientation": "portrait"
            }
          },
          "forceRegenerate": {
            "type": "boolean",
            "description": "Force regeneration even if PDF already exists",
            "example": false,
            "default": false
          }
        }
      },
...
    }
  }
}